requirements.txt
Django==5.2.7
django-simpleui>=2024.2.1
mcstatus>=11.0.0
django-debug-toolbar>=4.2.0
django-ckeditor>=6.7.0
Pillow>=10.0.0
python-dotenv>=1.0.1

# 缓存和性能优化
django-redis>=5.4.0
redis>=5.0.0

# 图片处理和优化
django-imagekit>=5.0.0
pilkit>=3.0

# 表单美化
django-crispy-forms>=2.1
crispy-bootstrap5>=2.0.0

# 静态文件压缩
django-compressor>=4.4

# 开发工具
django-extensions>=3.2.3
Faker>=22.0.0

# API 支持（可选）
djangorestframework>=3.14.0

# 安全和 CORS
django-cors-headers>=4.3.1

# 清理上传文件
django-cleanup>=8.1.0

# 异步任务（预留）
celery>=5.3

# 数据库驱动（可选）
psycopg2-binary>=2.9; platform_system != 'Windows' or platform_machine != 'ARM64'
psycopg[binary]>=3.1
mysqlclient>=2.2; platform_system == 'Windows' or platform_system == 'Linux'
